vietapp-bot/requirements.txt
botbuilder-integration-aiohttp>=4.14.0
botbuilder-ai>=4.14.0
botbuilder-sangamam-cognitive-text-translate